// src/app/interfaces/game.ts
export interface GameRequest {
  id?: number;
  leagueId: string;
  results: GameResult[];
  players: GamePlayers[];
}

export interface GamePlayers {
  id: number;
  gameId?: number;
}

export interface GameResult {
  id?: number;
  rank: number;
  playerId: number;
  point: number;
  calcPoint?: number;
  gameId?: number;
}

export interface GameResponse {
  id: number;
  leagueId: string;
  createdAt: string;
  updatedAt: string;
  deletedAt: string;
  results: GameResultResponse[];
}

export interface GameResultResponse {
  id: number;
  createdAt: string;
  updatedAt: string;
  deletedAt: string;
  playerId: number;
  playerName: string;
  rank: number;
  point: number;
  calcPoint: number;
  gameId: number;
}